[add] create configmap, rbac, worker and launcher

+++ api/v1/mpijob_types.go
@@ -21,16 +21,13 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-// EDIT THIS FILE!  THIS IS SCAFFOLDING FOR YOU TO OWN!
-// NOTE: json tags are required.  Any new fields you add must have json tags for the fields to be serialized.
-
 // MPIJobSpec defines the desired state of MPIJob
 type MPIJobSpec struct {
 	LauncherTemplate v1.PodTemplateSpec `json:"launcherTemplate"`
 
 	WorkerTemplate v1.PodTemplateSpec `json:"workerTemplate"`
 
-	NumWorkers *int64 `json:"numWorkers"`
+	NumWorkers *int `json:"numWorkers"`
 }
 
 // MPIJobStatus defines the observed state of MPIJob
